requirements.txt
# AI-Based Tomato & Potato Disease Classification
# Requirements for Local Development
# Author: Peter Maina (136532)

# Deep Learning Framework
tensorflow==2.13.0
keras==2.13.1

# Computer Vision
opencv-python==4.8.0.76
Pillow==10.0.0

# Data Science & Analysis
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0

# Data Augmentation
imgaug==0.4.0

# Configuration & Utilities
pyyaml==6.0.1
tqdm==4.66.1

# Kaggle API (for dataset download)
kaggle==1.5.16

# Jupyter Notebooks
jupyter==1.0.0
notebook==7.0.3
ipykernel==6.25.2

# Model Visualization
graphviz==0.20.1
pydot==1.4.2

# API Development (Phase 2 - Backend)
fastapi==0.103.0
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==2.3.0

# Database (Phase 2 - Backend)
pymongo==4.5.0
motor==3.3.1

# Authentication & Security (Phase 2)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
flake8==6.1.0
black==23.9.1
isort==5.12.0
pylint==2.17.5

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Deployment
gunicorn==21.2.0

# Monitoring & Logging
tensorboard==2.13.0

# Additional Utilities
requests==2.31.0
